pyproject.toml
[project]
name = "mdm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-sequences>=3.0",
]
